fix: can't call function in conditions block
